pyproject.toml
[project]
name = "withyou-chat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dspy>=3.0.3",
    "groq>=0.31.1",
    "ipykernel>=6.30.1",
    "mem0ai>=0.1.117",
    "ollama>=0.5.3",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.1.0",
    "streamlit>=1.49.1",
]
